4.21

+++ part4/bloglist/controllers/blogs.js
@@ -7,14 +7,6 @@ const User = require("../models/user");
 
 // Try/catch no escritos por la biblioteca "express-async-errors"
 
-const getTokenFrom = (request) => {
-  const authorization = request.get("authorization");
-  if (authorization && authorization.toLowerCase().startsWith("bearer ")) {
-    return authorization.substring(7);
-  }
-  return null;
-};
-
 blogsRouter.get("/", async (request, response) => {
   const blogs = await Blog.find({}).populate("user", {
     username: 1,
@@ -26,12 +18,8 @@ blogsRouter.get("/", async (request, response) => {
 
 blogsRouter.post("/", async (request, response) => {
   const body = request.body;
-
-  const token = getTokenFrom(request);
+  const token = request.token;
   const decodedToken = jwt.verify(token, process.env.SECRET);
-  if (!token || !decodedToken.id) {
-    return response.status(401).json({ error: "token missing or invalid" });
-  }
   const user = await User.findById(decodedToken.id);
 
   const blog = new Blog({
@@ -49,7 +37,7 @@ blogsRouter.post("/", async (request, response) => {
   }
 
   const savedBlog = await blog.save();
-  logger.info(`blog ${blog.title} successfully saved`);
+  logger.info(`blog "${blog.title}" successfully saved`);
 
   user.blogs = user.blogs.concat(savedBlog._id);
   await user.save();
@@ -91,9 +79,19 @@ blogsRouter.put("/:id", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id);
-  logger.info("blog successfully deleted");
-  response.status(204).end();
+  const token = request.token;
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+
+  const user = await User.findById(decodedToken.id);
+  const blogToDelete = await Blog.findById(request.params.id);
+
+  if (blogToDelete.user._id.toString() === user._id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id);
+    logger.info("blog successfully deleted");
+    response.status(204).end();
+  } else {
+    return response.status(401).json({ error: "user not authorized" });
+  }
 });
 
 module.exports = blogsRouter;
